change server dir
